// src/dto/driver/get.driver.dto.ts
export class GetDriverDto {
    driverId: number;
    agriculturalHoldingId: number;
    driverName: string;
    driverSurname: string;
    driverAddress: string;
    driverCity: string;
    zipCode: string;
    driverCardNum: string;
    driverPhone: string;
    driverEmail: string;
    driverStatus: boolean;
}